src/Jiro.Communication/requirements.txt
grequests==0.6.0
